Java - Design Pattern - Builder Pattern
Software that can be used by the attendee of a Design Patterns
Conference at the University. Attendee will register for the conference,
visit one of 3 labs to work on a programming contest, decide which of
the 3 presentation sessions to attend and decide which of the 3
restaurants to visit for a working-dinner. The attendee's choices will
result in a different time (in minutes), cost (in dollars), and
carbon-footprint.